// src/Pages/Modules/pos/order/Order.type.ts
interface DataType {
    key: string;
    orderNumber: string;
    date: string;
    quotationNumber: string;
    customer: string;
    subTotal: string;
    totalTax: string;
    grandTotal: string;
    orderStatus: string;
    deliveryStatus: string;
    invoiceStatus: string;
}

export type { DataType };